requirements.txt
langchain
langchain-core
langchain-community
langchain-ollama
pypdf
pymupdf
sentence-transformers
faiss-cpu
chromadb
python-dotenv
fastapi
uvicorn[standard]
pydantic[email]
python-multipart